requirements.txt
gunicorn==19.3.0
argcomplete==1.10.3
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blis==0.7.6
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
compressed-rtf==1.0.6
cryptography==36.0.1
cymem==2.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
Flask==2.0.3
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth-httplib2==0.1.0
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
IMAPClient==2.1.0
importlib-resources==5.4.0
install==1.3.5
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.0
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
olefile==0.46
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
parse==1.19.0
pathy==0.6.1
pdf2image==1.16.0
pdfminer.six==20191110
Pillow==9.0.1
pip==22.0.4
preshed==3.0.6
protobuf==3.19.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.8.2
PyMuPDF==1.19.6
pyparsing==3.0.7
PyPDF2==1.26.0
pyresparser==1.0.6
pyrsistent==0.18.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
regex==2022.3.15
requests==2.27.1
rsa==4.8
setuptools==60.9.3
six==1.12.0
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.1
spacy-lookups-data==1.0.3
spacy==2.2.4
SpeechRecognition==3.8.1
srsly==2.4.2
textract==1.6.5
thinc==8.0.14
tqdm==4.63.0
typer==0.4.0
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
wasabi==0.9.0
Werkzeug==2.0.3
xlrd==1.2.0
XlsxWriter==3.0.3
zipp==3.7.0
